Add configVolumeMountSubPath util function (#58)
* Add configVolumeMountSubPath util function

Gives user possibility of mounting configmap to a file in a directory rather than mounting the whole directory. e.g to /etc/config/config.file rather than to /etc/config/.

Is this something you would want? Would the better implementation be to have logic in configVolumeMount that uses subpath if there is a third argument (I'm not sure yet how to do that with jsonnet)?

* Implement optional volumeMountMixin with comment

* Add volumeMountMixin to *VolumeMount helper functions

* Remove copy-paste mistake