Add node-mixin to prometheus-ksonnet
There is a non-empty intersection between the kubernetes-mixin and the
node-mixin in terms of rules and dashboards. However, the names are
actually all different (kind of deliberate for the dashboards - for
the rules we are "lucky" that the kubernetes-mixin is totally not
following the Prometheus rule naming convetions, while the node-mixin
is by now).

So this should just play well along each other for now. Of course, we
want to deduplicate the more or less similar dashboards eventually. We
still have to figure out on which level this should happen. We either
kick out some dashboards here in prometheus-ksonnet or, probably
better, make upstream kubernetes-mixin use the node-mixin instead of
doing their own stuff.

Historically, I guess, the node-mixin parts that overlap with the
kubernetes-mixin were copied over from or at least inspired by the
kubernetes-mixin. It probably makes a lot of sense to let the
kubernetes-mixin use the node-mixin maintained in the
prometheus/node_exporter repo directly.

But let's first play with it and then make that call later.