Adjust to new node mixin version.
For one, this disables paging alerts for disks running full. The nodes
we monitor are K8s nodes. If individual notes kill themselves by
filling their disk, we don't need to wake somebody up. K8s will cope
with it. Hence, those alerts should all be just warnings.

The other change here is to kick out the node-related dashboards from
the k8s mixin. They are obsolete now. Much better versions are
contained in the node mixin. In fact, the current version of the
k8s-mixin has kicked those out upstream. However, the current version
of the k8s-mixin also requires K8s 1.14+ (which changed a lot of the
metrics names). Since we are stuck on 1.13 and before for a while, we
have to keep using an older version of the k8s-mixin.