Switch from DNS-SRV to plain DNS for non-local Alertmanagers
I assumed, the K8s-provided DNS-SRV records for named ports would
resolve into all the individual pods having that port. Instead, it
resolves to the load-balanced service IP number. That doesn't help us
because Prometheus servers have to send their alerts to _each_
Alertmanager, not just one of them.

So we have to hardwire the DNS names of each AM pod. The only downside
is that a change of the replica count is not autopropagated but
creates a `tk diff` that needs to be applied explicitly.