Enable Prometheus TSDB admin API by default
I think this is useful to have on by default. We can take snapshot if
we need to, or cleanup tombstones. We can also delete individual
series, which is at the same time the risk (if it happens
accidentally). However, since we remote-write everything to Cortex, I
think the risk is quite limited.

But perhaps there was another reason to not enable it?

Details about the API see
https://prometheus.io/docs/prometheus/latest/querying/api/#tsdb-admin-apis