Add __replica__ labels to prometheus-ha replicas
Those labels are needed by Cortex for deduplication. In our existing
setup, we use "one" and "two" as labels. Here I have chosen "zero" and
"one", in line with the zero-based numbering within a stateful set. I
expect that this should not make any difference for Cortex. It only
matters that both label values are different.