Remove "mixin" from prometheus-ksonnet title (#298)
As per definition (https://en.wikipedia.org/wiki/Mixin), a mixin is simply spoken a piece of code that's mixed with another existing piece of code.

This does apply to monitoring mixins, because they are mixed into this very library. prometheus-ksonnet however itself is not a mixin and should thus not be titled as one.