Configure securityContext for AlertManager statefulset.
AlertManager became a non-root image some time ago. Since in StatefulSet
variant it writes some state to an external PV, which is usually
ext4-formatted and root-owned, this prevents AM from writing to the PV.

Prometheus had a similar issue which has been resolved by adding
`securityContext` to the StatefulSet in
https://github.com/grafana/jsonnet-libs/pull/244; let's do the same for
AlertManager.

Signed-off-by: Milan Plzik <milan.plzik@grafana.com>