Update weaveworks/watch to the current version
This version fixes a bug that causes misses of some updates,
especially when whole directories are being watched in the context of
K8s configmaps.

Note that the upstream weaveworks/Watch repo currently has trouble
with their CI pipeline. So I built the image from the same commit
myself and pushed it to my own Docker Hub account. Once the same image
is also on the Weaveworks account, we should switch back.

Signed-off-by: beorn7 <beorn@grafana.com>