memcached: Ensure container memory request is <= limit
Currently for small memory_limit_mb values, (eg 128), the container
resource request will end up larger than the limit, which is invalid.

To handle this, take the smaller of the calculated memory request
value and the memory limit as the request, ensuring even for small
values, that the request is set to the limit.