Add jsonnetfile for the envoy mixin and fix other linting errors.
This fixes the `mixin lint` step with more recent versions of the
`mixtool` tool. The problem is likely not appearing in CI as the
`mixtool` version used is a few months old.

I've updated the Dockerfile as well to use the latest available revision
of `mixtool`, so the a new docker image should be published once this is
merged.