refactor(etcd-operator): Remove kausal from root element
Instead, declare it explicitly, because it is an implementation detail
that is leaking to the consumers of the libraries, polluting the root
element.

Particularly, this solves one use case, which is importing
`etcd_cluster`:

```
local etcd_cluster = import 'etcd-operator/etcd-cluster.libsonnet';

{
  etcd_cluster: etcd_cluster.etcd_cluster('etcd')
}
```

won't work if kausal is not available on the root element.

To solve that we would need to do:

```
local etcd_cluster = import 'etcd-operator/etcd-cluster.libsonnet';
local k = import 'ksonnet-util/kausal.libsonnet';

{
  etcd_cluster: (k + etcd_cluster).etcd_cluster('etcd')
}
```

For what is worth, that behaviour was not present when importing
'etcd-operator/etcd-operator.libsonnet'.