Allow smaller memcached deployments
Sometimes in dev env we want to deploy some small memcached instances,
with memory limits like 64Mb. However, the calculation for
memory_request_bytes adds an extra 100Mb overhead to the requested
memory but not to the memory limit, which makes k8s fail to schedule the
StatefulSet.

This changes two things: we ensure that memory_limit_bytes is at least
as big as memory_request_bytes, plus we make the the requested memory
overhead configurable.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>