ksonnet-util: Check mapToFlags input
Golang accepts both -foo and --foo flags, and sometimes jsonnet can
become complex and there can be a 'foo' option and a '-foo' override
somewhere else, causing both flags to be rendered. I don't think that's
a desired behavior for anyone, so added a check by default, that can be
disabled if needed.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>