nginx-directory: allow disabling downstream subfilter
There's a `subfilter` property that allows rewriting request paths.
Sometimes, several nginx-directory instances are chained, and if first
has subfilter enabled, then the next ones have no control over what
happens to the HTML they return.

This might be undesirable since subfilter doesn't always cover all the
cases, and some apps just want to know the data they return.

This adds a `disable_downstream_subfilter` option to the route, which
makes the directory proxy_pass to add `X-Nginx-Subfilter: disable` to
all responses. On the other hand, the subfilter rules now check for that
header, and if it's set to `disable` then subfilter is not applied.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>