prometheus/ha-mixin: Add a startup probe
Prometheus replays the WAL on startup, during which time it's not ready
for requests. This can take a long time for large environments, longer
than the period the readiness probe considers to be OK. So we give a
generous 25 minutes for this to happen. This can be tuned up or down by
consumers of course.