Fixes a number of issues to Solr Mixin.
Fixes issue with Resource Monitoring Dashboard where '~' was missing in query.
Change custom allValues for solr_collection variable in Cluster Overview Dashboard
which was causing an issue in panels when collection label was not present in metrics.
Change custom allValues for solr_collection variable in Query Performance Dashboard
which was causing issue in panels when collection label was not present in metrics.