nginx-directory: add support for non-normalized proxying & non-absolute links (#1501)
By default this library builds the links starting with a slash, which makes them absolute so they only work if nginx serves the root, or if it's being proxied by another nginx with a subfilter directive.

The subfilter directive however isn't as awesome as it sounds: it can't catch it all (easily misses AJAX urls) and it's very awkward to use when proxying through one nginx with subfilter and another one without it.

In a similar way, the normalization _magic_ that nginx performs when providing an URL to proxy_pass is sometimes unexpected: it's easier to just expect the URL that has been provided to nginx, even when that means that the proxied service must know the full path from root.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>